# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/ai/agents": {
      "post": {
        "summary": "Create AI agent with immutable version 1",
        "security": [
          {
            "bearer": []
          },
          {
            "tenant": []
          }
        ],
        "tags": [
          "AI"
        ]
      },
      "get": {
        "summary": "List AI agents",
        "security": [
          {
            "bearer": []
          },
          {
            "tenant": []
          }
        ],
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/agents/{id}": {
      "get": {
        "summary": "Get AI agent",
        "security": [
          {
            "bearer": []
          },
          {
            "tenant": []
          }
        ],
        "tags": [
          "AI"
        ]
      },
      "patch": {
        "summary": "Update AI agent (creates new version when config changes)",
        "security": [
          {
            "bearer": []
          },
          {
            "tenant": []
          }
        ],
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/agents/{id}/activate": {
      "post": {
        "summary": "Activate AI agent route",
        "security": [
          {
            "bearer": []
          },
          {
            "tenant": []
          }
        ],
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/agents/{id}/disable": {
      "post": {
        "summary": "Disable AI agent",
        "security": [
          {
            "bearer": []
          },
          {
            "tenant": []
          }
        ],
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/ai/agents/{id}/versions": {
      "get": {
        "summary": "List immutable AI agent versions",
        "security": [
          {
            "bearer": []
          },
          {
            "tenant": []
          }
        ],
        "tags": [
          "AI"
        ]
      }
    },
...
  },
  "info": {
    "title": "PBX Platform API",
    "description": "Multi-tenant virtual PBX control plane API. Tenant-scoped routes require Bearer JWT and X-Tenant-Id header.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "tenant": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-Id"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Tenant API key: Authorization: Bearer pbx_live_<prefix>_<secret>"
      }
    },
...
  }
}